+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/animation/progress/ProgressBarAnimation.kt
@@ -1,7 +1,6 @@
 package com.github.ajalt.mordant.animation.progress
 
 import com.github.ajalt.mordant.animation.Refreshable
-import com.github.ajalt.mordant.widgets.progress.ProgressBarCell
 import com.github.ajalt.mordant.widgets.progress.ProgressBarDefinition
 import com.github.ajalt.mordant.widgets.progress.ProgressState
 import com.github.ajalt.mordant.widgets.progress.TaskId