+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/terminal/TerminalColors.kt
@@ -14,7 +14,7 @@ import com.github.ajalt.mordant.rendering.Theme
  * Strings decorated with these styles will be downsampled correctly even if they are printed
  * directly rather than through [Terminal.print].
  */
-class TerminalColors internal constructor(
+class TerminalColors internal constructor( // TODO(3.0): remove
     private val terminalInfo: TerminalInfo,
     private val theme: Theme,
 ) {