+++ mordant/src/jsMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.js.kt
@@ -1,11 +1,7 @@
 package com.github.ajalt.mordant.internal
 
-import com.github.ajalt.mordant.internal.syscalls.SyscallHandlerJsCommon
-import com.github.ajalt.mordant.internal.syscalls.SyscallHandlerJsNode
-import com.github.ajalt.mordant.internal.syscalls.SyscallHandlerNode
-import com.github.ajalt.mordant.terminal.PrintTerminalCursor
-import com.github.ajalt.mordant.terminal.Terminal
-import com.github.ajalt.mordant.terminal.TerminalCursor
+import com.github.ajalt.mordant.terminal.`interface`.TerminalInterfaceJsCommon
+import com.github.ajalt.mordant.terminal.`interface`.TerminalInterfaceJsNode
 
 private external val process: dynamic
 private external val console: dynamic
@@ -17,9 +13,9 @@ internal actual fun browserPrintln(message: String) {
     console.error(message)
 }
 
-internal actual fun makeNodeSyscallHandler(): SyscallHandlerJsCommon? {
+internal actual fun makeNodeTerminalInterface(): TerminalInterfaceJsCommon? {
     return try {
-        SyscallHandlerJsNode(nodeRequire("fs"))
+        TerminalInterfaceJsNode(nodeRequire("fs"))
     } catch (e: Exception) {
         null
     }