+++ mordant/src/posixMain/kotlin/com/github/ajalt/mordant/internal/tty.kt
@@ -1,6 +1,6 @@
 package com.github.ajalt.mordant.internal
 
-import com.github.ajalt.mordant.internal.syscalls.SyscallHandlerPosix
+import com.github.ajalt.mordant.internal.syscalls.SyscallHandlerNativePosix
 import platform.posix.ECHO
 
 // https://www.gnu.org/software/libc/manual/html_node/getpass.html